Refactor shared setting component
