+++ rightnow-firebase/src/components/share_settings/user_change_password.js
