+++ rightnow-firebase/src/components/share_settings/user_notification.js
